+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+* [Qiniu] Custom domain and secure connection.
+
 ## [v0.4.0](https://github.com/linjunpop/blast_off/tree/v0.4.0)
 
 * Generate upload token based on bucket and key.